+++ CHANGES.rst
@@ -9,6 +9,7 @@ Improve the situation with large Grafana installations, see #2.
 Thanks `@jangaraj <https://github.com/jangaraj>`_.
 
 - Add option ``--cache-ttl`` for controlling the cache expiration time
+- Improve error logging when hitting Grafana unauthorized
 
 
 2019-04-21 0.3.1